a ver si así va la query

+++ src/main/java/repositories/ChorbiRepository.java
@@ -92,10 +92,10 @@ public interface ChorbiRepository extends JpaRepository<Chorbi, Integer> {
 
 	//B1: The minimum, the maximum, and the average number of stars per chorbi.
 	@Query("select sum(s.stars) from Sense s group by s.recipient order by sum(s.stars) ASC")
-	Double[] minStarsPerChorbi();
+	Integer[] minStarsPerChorbi();
 
 	@Query("select sum(s.stars) from Sense s group by s.recipient order by sum(s.stars) DESC")
-	Double[] maxStarsPerChorbi();
+	Integer[] maxStarsPerChorbi();
 
 	@Query("select sum(s.stars)/(select count(c1) from Chorbi c1)*1.0 from Sense s")
 	Double avgStarsPerChorbi();
